refactor: extract RunbookActorService and add asTerminalSnapshot type guard
Extract actor lifecycle logic (create, sync, sendAndSync) from
RunbookStateManager into dedicated RunbookActorService class.
ForIterationService now depends on narrow interfaces (ForStateReader,
ForActorOperations) instead of concrete classes.

Add asTerminalSnapshot() to snapshot-utils.ts, replacing unsafe bare
casts and verbose inline type narrowing across execution.ts,
for-iteration-service.ts, and transitions.ts.
