+++ packages/core/src/runbook/actor-service.ts
@@ -1,0 +1,244 @@
+// packages/core/src/runbook/actor-service.ts
+
+/**
+ * XState actor lifecycle service for runbooks.
+ *
+ * Owns actor creation (with snapshot migration), state synchronisation
+ * after transitions, and convenience methods for the two dominant patterns:
+ * initialisation (create + sync, no event) and transition (create + send + sync).
+ *
+ * Composes RunbookStateManager for persistence — does not own disk I/O.
+ *
+ * @module
+ */
+
+import { createActor, type AnyActorRef } from 'xstate';
+import type { Step, RunbookState, ForContext } from './types.js';
+import type { RunbookStateManager } from './state.js';
+import { compileRunbookToMachine, type RunbookEvent } from './compiler.js';
+
+// Re-export so CLI callers import from core, not directly from xstate
+export type { AnyActorRef } from 'xstate';
+
+/**
+ * Result of a {@link RunbookActorService.sendAndSync} operation.
+ *
+ * Bundles the actor, updated persisted state, and raw snapshot so callers
+ * can inspect terminal states (COMPLETE / STOPPED) without an extra call.
+ */
+export interface ActorSyncResult {
+  /** The XState actor (started, event already sent) */
+  actor: AnyActorRef;
+  /** The persisted RunbookState after syncing from actor snapshot */
+  state: RunbookState;
+  /** The raw persisted snapshot for terminal-state inspection */
+  snapshot: unknown;
+}
+
+/**
+ * Manages XState actor lifecycle for runbooks.
+ *
+ * Encapsulates actor creation (with snapshot migration), state synchronisation
+ * after transitions, and convenience methods for the two dominant usage patterns:
+ * initialisation (create + sync with no event) and transition (create + send + sync).
+ */
+export class RunbookActorService {
+  constructor(private readonly manager: RunbookStateManager) {}
+
+  /**
+   * Create and start an XState actor from persisted state.
+   *
+   * Loads the persisted snapshot, applies migration from flat FOR fields
+   * to forStack array if needed, compiles the machine, and starts the actor.
+   *
+   * @param id - Runbook state ID
+   * @param steps - Parsed runbook steps for machine compilation
+   * @returns Started actor, or null if state not found
+   */
+  async createActor(id: string, steps: Step[]): Promise<AnyActorRef | null> {
+    const state = await this.manager.load(id);
+    if (!state) return null;
+
+    const machine = compileRunbookToMachine(steps, { sources: state.sources });
+
+    // Migrate old snapshot context: flat FOR fields → forStack
+    // Snapshot migration deals with untyped persisted data — any is unavoidable
+    /* eslint-disable @typescript-eslint/no-explicit-any, @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-unsafe-member-access, @typescript-eslint/no-unnecessary-type-assertion */
+    const snapshot = state.snapshot as any;
+    if (snapshot?.context && !snapshot.context.forStack) {
+      const ctx = snapshot.context as any;
+      if (ctx.forIteration !== undefined) {
+        // Derive stepId from snapshot.value (authoritative) with state.step fallback
+        const stateValue = snapshot.value as string | undefined;
+        const stepMatch = stateValue
+          ? (/^step::([^:]+)/.exec(stateValue) ?? /^step_([^_]+)/.exec(stateValue))
+          : null;
+        const stepId = stepMatch?.[1] ?? state.step;
+
+        snapshot.context = {
+          ...ctx,
+          forStack: [
+            {
+              stepId,
+              iteration: ctx.forIteration,
+              start: ctx.forStart ?? 1,
+              end: ctx.forEnd ?? ctx.forIteration,
+              variable: ctx.forVariable,
+              source: { kind: 'range' as const },
+            },
+          ],
+          forIteration: undefined,
+          forStart: undefined,
+          forEnd: undefined,
+          forVariable: undefined,
+        };
+      } else {
+        // No active loop — just ensure forStack exists
+        snapshot.context = { ...ctx, forStack: [] };
+      }
+    }
+    /* eslint-enable @typescript-eslint/no-explicit-any, @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-unsafe-member-access, @typescript-eslint/no-unnecessary-type-assertion */
+
+    const actor = createActor(machine, {
+      // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
+      snapshot,
+    });
+    actor.start();
+    return actor;
+  }
+
+  /**
+   * Synchronise persisted state from actor snapshot.
+   *
+   * Extracts step/substep position, variables, forStack, and lastAction
+   * from the actor's persisted snapshot and writes to disk.
+   *
+   * @param id - Runbook state ID
+   * @param actor - The XState actor to read snapshot from
+   * @param steps - Parsed runbook steps for step name lookup
+   * @returns Updated persisted RunbookState
+   */
+  async updateFromActor(id: string, actor: AnyActorRef, steps: Step[]): Promise<RunbookState> {
+    // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-explicit-any
+    const snapshot = actor.getPersistedSnapshot() as any;
+    // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
+    const stateValue = snapshot.value as string;
+
+    // If the runbook is in a final state, don't try to parse a step number.
+    // Just update the snapshot and variables, preserving the last step number.
+    if (stateValue === 'COMPLETE' || stateValue === 'STOPPED') {
+      // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
+      const variables = (snapshot.context?.variables ?? {}) as Record<
+        string,
+        boolean | number | string
+      >;
+      return await this.manager.update(id, {
+        variables,
+        snapshot,
+        // Clear FOR loop state on completion
+        forStack: undefined,
+        iterationResults: undefined,
+      });
+    }
+
+    // Parse step name from XState state value
+    const primaryMatch = /^step::(.+?)(?:::(.+))?$/.exec(stateValue);
+    const legacyMatch = !primaryMatch ? /^step_([^_]+)(?:_([^_]+))?$/.exec(stateValue) : null;
+    if (legacyMatch) {
+      console.warn(
+        'Deprecated state-ID format "step_…" detected. Please restart execution to migrate to "step::…" format.',
+      );
+    }
+    const match = primaryMatch ?? legacyMatch;
+    const stepName = match ? match[1] : steps[0].name;
+
+    // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
+    let substep = snapshot.context?.substep as string | undefined;
+    if (!substep && match?.[2]) {
+      substep = match[2];
+    }
+
+    // Find step by name (unified lookup)
+    const step = steps.find((s) => s.name === stepName) ?? steps[0];
+
+    // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
+    const retryCount = snapshot.context?.retryCount as number;
+    // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
+    const variables = (snapshot.context?.variables ?? {}) as Record<
+      string,
+      boolean | number | string
+    >;
+
+    // FOR loop context
+    // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
+    const forStack = snapshot.context?.forStack as ForContext[] | undefined;
+    // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
+    const iterationResults = snapshot.context?.iterationResults as ('pass' | 'fail')[] | undefined;
+    // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
+    const lastAction = snapshot.context?.lastAction as RunbookState['lastAction'];
+
+    return await this.manager.update(id, {
+      step: stepName, // string
+      substep,
+      stepName: step.description,
+      retryCount,
+      variables,
+      snapshot,
+      // Filter implicit ForContext entries — don't persist synthetic loop state
+      forStack: (() => {
+        const realForStack = forStack?.filter((fc) => !fc.implicit);
+        return realForStack?.length ? realForStack : undefined;
+      })(),
+      // Only clear iterationResults when all stack entries were implicit.
+      // When forStack is empty after explicit FOR exit, iterationResults
+      // must be preserved for parent-step aggregation.
+      iterationResults: (() => {
+        const hasOnlyImplicit = forStack?.length ? forStack.every((fc) => fc.implicit) : false;
+        return hasOnlyImplicit ? undefined : iterationResults;
+      })(),
+      lastAction,
+    });
+  }
+
+  /**
+   * Initialise actor state without sending any event.
+   *
+   * Creates actor, syncs snapshot to persistence, and returns the state.
+   * Used at runbook startup to populate initial forStack/context.
+   *
+   * @param id - Runbook state ID
+   * @param steps - Parsed runbook steps
+   * @returns Updated state, or null if state not found
+   */
+  async initializeState(id: string, steps: Step[]): Promise<RunbookState | null> {
+    const actor = await this.createActor(id, steps);
+    if (!actor) return null;
+    return await this.updateFromActor(id, actor, steps);
+  }
+
+  /**
+   * Create actor, send event, sync state, and return all three artefacts.
+   *
+   * This is the dominant usage pattern: create actor from persisted state,
+   * send a transition event (PASS/FAIL/GOTO), sync the result back to disk,
+   * and return the actor + state + snapshot for the caller to inspect
+   * terminal states and extract context.
+   *
+   * @param id - Runbook state ID
+   * @param steps - Parsed runbook steps
+   * @param event - Runbook event to send (PASS, FAIL, RETRY, or GOTO)
+   * @returns Actor, updated state, and snapshot; or null if state not found
+   */
+  async sendAndSync(
+    id: string,
+    steps: Step[],
+    event: RunbookEvent,
+  ): Promise<ActorSyncResult | null> {
+    const actor = await this.createActor(id, steps);
+    if (!actor) return null;
+    actor.send(event);
+    const state = await this.updateFromActor(id, actor, steps);
+    const snapshot = actor.getPersistedSnapshot();
+    return { actor, state, snapshot };
+  }
+}
